% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbstgrid.r
\docType{package}
\name{rbstgrid}
\alias{rbstgrid}
\title{An R Package to Aggregate Grid Data to a Different Cell Size}
\description{
This package includes a function to aggregate grid data to a different cell size.
}
\details{
This package includes a function to aggregate grid data to a different cell size.

For a theoretical rationale and an empirical example, please see the following paper:

Suzuki, Akisato. 2022. "Uncertainty in Grid Data: A Theory and Comprehensive
Robustness Test." arXiv:2202.03241 [stat.ME]. https://arxiv.org/abs/2202.03241.

If you use this package, please cite the following items:

Suzuki, Akisato. 2022. "Uncertainty in Grid Data: A Theory and Comprehensive
Robustness Test." arXiv:2202.03241 [stat.ME]. https://arxiv.org/abs/2202.03241.

Elkink, Johan A., and Akisato Suzuki. 2022. "rbstgrid: An R Package to
Aggregate Grid Data to a Different Cell Size." R package version 0.0.0.9001.
}
\section{Authors}{

Author: Johan A. Elkink (\email{jos.elkink@ucd.ie})

Author & Maintainer: Akisato Suzuki (\email{akisato.suzuki@gmail.com})
}

\section{rbstgrid functions}{

aggr_grid: A function to aggregate grid data to a different cell size
}

