% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbstgrid.r
\docType{package}
\name{rbstgrid}
\alias{rbstgrid}
\title{An R Package to Aggregate Grid Data to a Different Cell Size}
\description{
This package includes a function to aggregate grid data to a different cell size.
}
\details{
This package includes a function to aggregate grid data to a different cell size.

For a theoretical rationale and an empirical example, please see the
following paper:

Suzuki, Akisato. 2022. "Uncertainty in Grid Data: A Theory and Comprehensive
Robustness Test." Quality & Quantity. https://doi.org/10.1007/s11135-022-01555-x.

If you use this package, please cite the following items:

Suzuki, Akisato. 2022. "Uncertainty in Grid Data: A Theory and Comprehensive
Robustness Test." Quality & Quantity. https://doi.org/10.1007/s11135-022-01555-x.

Dornschneider-Elkink, Johan A., and Akisato Suzuki. 2022. "rbstgrid: An R
Package to Aggregate Grid Data to a Different Cell Size." R package version 1.0.0. https://github.com/AkisatoSuzuki/rbstgrid.
}
\section{Authors}{

Author: Johan A. Dornschneider-Elkink (\email{jos.elkink@ucd.ie})

Author & Maintainer: Akisato Suzuki (\email{akisato.suzuki@gmail.com})
}

\section{rbstgrid functions}{

aggr_grid: A function to aggregate grid data to a different cell size
}

